requirements.txt
annotated-types==0.7.0
anyio==4.5.0
certifi==2024.8.30
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-telegram-bot==21.6
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
